Move files into lib folder
