+++ lib/gitignore.js
