+++ lib/options-manager.js
@@ -36,8 +36,8 @@ const DEFAULT_CONFIG = {
 	baseConfig: {
 		extends: [
 			'xo',
-			path.join(__dirname, 'config/overrides.js'),
-			path.join(__dirname, 'config/plugins.js')
+			path.join(__dirname, '../config/overrides.js'),
+			path.join(__dirname, '../config/plugins.js')
 		]
 	}
 };
@@ -124,7 +124,7 @@ const buildConfig = opts => {
 	if (opts.esnext !== false) {
 		config.baseConfig.extends = [
 			'xo/esnext',
-			path.join(__dirname, 'config/plugins.js')
+			path.join(__dirname, '../config/plugins.js')
 		];
 	}
