+++ package.json
@@ -30,11 +30,9 @@
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "files": [
-    "gitignore-filter.js",
-    "index.js",
-    "cli.js",
-    "options-manager.js",
-    "config"
+    "config",
+    "lib",
+    "*.js"
   ],
   "keywords": [
     "❤️",
